chore: sync latest changes from origin

+++ apps/web/src/app/api/meet-together/plans/[planId]/route.ts
@@ -1,26 +1,29 @@
-import { createRouteHandlerClient } from '@supabase/auth-helpers-nextjs';
 import { NextResponse } from 'next/server';
-import { cookies } from 'next/headers';
+import { createAdminClient } from '@/utils/supabase/client';
 
 export const dynamic = 'force-dynamic';
 
 interface Params {
   params: {
-    timezoneId: string;
+    planId: string;
   };
 }
 
-export async function PUT(
-  req: Request,
-  { params: { timezoneId: id } }: Params
-) {
-  const supabase = createRouteHandlerClient({ cookies });
+export async function PUT(req: Request, { params: { planId: id } }: Params) {
+  const sbAdmin = createAdminClient();
+
+  if (!sbAdmin)
+    return NextResponse.json(
+      { message: 'Internal server error' },
+      { status: 500 }
+    );
 
   const data = await req.json();
+  const name = data.name;
 
-  const { error } = await supabase
+  const { error } = await sbAdmin
     .from('meet_together_plans')
-    .upsert(data)
+    .update({ name })
     .eq('id', id);
 
   if (error) {
@@ -34,13 +37,16 @@ export async function PUT(
   return NextResponse.json({ message: 'success' });
 }
 
-export async function DELETE(
-  _: Request,
-  { params: { timezoneId: id } }: Params
-) {
-  const supabase = createRouteHandlerClient({ cookies });
+export async function DELETE(_: Request, { params: { planId: id } }: Params) {
+  const sbAdmin = createAdminClient();
 
-  const { error } = await supabase
+  if (!sbAdmin)
+    return NextResponse.json(
+      { message: 'Internal server error' },
+      { status: 500 }
+    );
+
+  const { error } = await sbAdmin
     .from('meet_together_plans')
     .delete()
     .eq('id', id);
